refactor: move these methods into iroh::util for now
