+++ iroh/src/lib.rs
@@ -6,3 +6,4 @@ pub use iroh_net as net;
 
 pub mod node;
 pub mod rpc_protocol;
+pub mod util;
